+++ SuperBuild/External_STIR.cmake
@@ -151,6 +151,7 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
        ${Boost_CMAKE_ARGS}
        ${HDF5_CMAKE_ARGS}
        ${${proj}_EXTRA_CMAKE_ARGS}
+    ${HDF5_EP_ARGS}
     DEPENDS
         ${${proj}_DEPENDENCIES}
   )